backend_refernce/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Vector Database
chromadb==0.4.18

# AI/ML
google-generativeai==0.3.2
openai==1.3.7

# YouTube API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.1

# Image processing
Pillow==10.1.0
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
